requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict==0.8.3
torchdata
transformers
wandb
packaging>=20.0
uvicorn
fastapi
latex2sympy2_extended
math_verify[antlr4_9_3]
vllm==0.10.0

# requirements for problem synthesis
sentence_transformers==3.2.1
scikit-learn==1.3.2
scipy==1.10.1
faiss-gpu==1.7.2
vllm==0.6.3
fire==0.7.0
str2bool